refactor: extract startup catch-up module
